Move JweEncryptedSerializer to seperate file
